+++ app/Components/ISharerLink.php
@@ -6,5 +6,6 @@ namespace Nexendrie\Components;
 interface ISharerLink {
   public function getShareLink(string $url): string;
   public function getSiteName(): string;
+  public function getPlatformName(): string;
 }
 ?>